Refactor code to add on concern
